Pull request #4010: [BLUETOOTH-17598] APP: TPMS additional MISRA fixes
Merge in CONNINT/bluetooth_public from BLUETOOTH-17598-misra to develop/main

Squashed commit of the following:

commit 28656098a1b582a12e678b504365926e88b22a75
Author: Silviu Petria <silviu.petria@nxp.com>
Date:   Mon May 11 10:16:26 2026 +0300

    [BLUETOOTH-17598] APP: TPMS additional MISRA fixes

+++ examples/tpms_monitor/tpms_monitor.c
@@ -1155,14 +1155,14 @@ static void ProcessNonConnScanEvent(gapScannedDevice_t* pData)
     uint8_t tpmsSigningKey[16] = {0U};
     uint8_t keyReversed[16] = {0U};
     bool_t bFound = FALSE;
-    uint8_t index = 0;
+    uint32_t index = 0;
     uint8_t optionalFieldsLen = 0U;
 
     while (index < pData->dataLength)
     {
         adElement.length = pData->data[index];
-        adElement.adType = (gapAdType_t)pData->data[index + 1];
-        adElement.aData = &pData->data[index + 2];
+        adElement.adType = (gapAdType_t)pData->data[index + 1U];
+        adElement.aData = &pData->data[index + 2U];
 
         if(adElement.adType == gAdServiceData16bit_c)
         {
